Add a module for the data landing zone
