+++ aleph/manage.py
@@ -1,7 +1,6 @@
 # coding: utf-8
 import os
 import logging
-import time
 from banal import ensure_list
 from normality import slugify
 from ingestors.util import decode_path
@@ -9,13 +8,13 @@ from alephclient.util import load_config_file
 from flask_script import Manager, commands as flask_script_commands
 from flask_script.commands import ShowUrls
 from flask_migrate import MigrateCommand
-from sqlalchemy.exc import DBAPIError
 
-from aleph.core import create_app, db, cache, get_dataset
-from aleph.model import Collection, Document, DocumentRecord, Role
+from aleph.core import create_app, db, cache
+from aleph.model import Collection, Document, Role
 from aleph.migration import upgrade_system, destroy_db, cleanup_deleted
 from aleph.views import mount_app_blueprints
 from aleph.index.admin import delete_index
+from aleph.logic.landing import export_balkhash_collection
 from aleph.logic.collections import create_collection, update_collection
 from aleph.logic.collections import index_collections, index_collection
 from aleph.logic.collections import delete_collection, delete_bulk_entities
@@ -189,54 +188,13 @@ def rdf(foreign_id):
             print(line)
 
 
-def _export_balkhash_collection(collection, retries=0, backoff=30, offset=0):
-    MAX_RETRIES = 5
-    RETRY_BACKOFF_FACTOR = 2
-    try:
-        from followthemoney import model
-        dataset = get_dataset(collection.foreign_id)
-        writer = dataset.bulk()
-        q = Document.by_collection(collection.id)
-        q = q.order_by(Document.id.asc()).offset(offset)
-        for doc in q.yield_per(5000):
-            log.debug("Export [%s:%s]: %s", doc.id, doc.schema, doc.name)
-            dproxy = doc.to_proxy()
-            writer.put(dproxy)
-            if doc.supports_records:
-                q = db.session.query(DocumentRecord)
-                q = q.filter(DocumentRecord.document_id == doc.id)
-                for record in q.yield_per(100):
-                    rproxy = record.to_proxy()
-                    writer.put(rproxy)
-                    dpart = model.make_entity(doc.schema)
-                    dpart.id = dproxy.id
-                    dpart.add('indexText', list(record.texts))
-                    writer.put(dpart, fragment=str(record.id))
-            offset += 1
-        dataset.close()
-    except DBAPIError as exc:
-        if retries < MAX_RETRIES:
-            log.debug("Error occurred: %s", exc)
-            log.debug("Retrying in %s seconds", backoff)
-            db.session.close()
-            dataset.close()
-            time.sleep(backoff)
-            retries = retries + 1
-            backoff = backoff * RETRY_BACKOFF_FACTOR
-            return _export_balkhash_collection(
-                collection, retries, backoff, offset
-            )
-        else:
-            log.exception(exc)
-
-
 @manager.command
 def exportbalkhash(foreign_id=None):
     collections = Collection.all()
     if foreign_id is not None:
         collections = [get_collection(foreign_id)]
     for collection in collections:
-        _export_balkhash_collection(collection)
+        export_balkhash_collection(collection)
 
 
 @manager.command
